Fix nondeterministic test
In one of the cache tests the test result depended on the order in which the queries were compiled.

This order is nondeterministic.

In case of a query with leading whitespace being compiled first, this led the test to expect a query where only whitespace that isn't part of the AST is found.

To fix this hardcoded positions are used instead.

Fixes #124

Signed-off-by: Tobias Guggenmos <tobias.guggenmos@uni-ulm.de>